requirements_cpu.txt
numpy
scipy
opencv-python
tensorflow-cpu==1.15
tqdm
python-benedict